fix: reference unique Textbox using set instead of list
Fixes #19.

When using PNG Info Send to txt2img/img2img, it will re-create the
entire tab for that workflow. This means that Textbox gets re-created.
On creation, the Textbox is added to the prompts to format, which makes
the old ones stale. Since there's excess, it also results in an uncaught
error.

Solution is to use a set instead of a list to hold the Textbox's to
format. I speculate the components have an internal check to ensure
they don't overlap, perhaps by checking `elem_id` or something.
Regardless, fixed.